⚛️ Fuse `myst-spec-ext` into `myst-spec` (#2908)
