+++ packages/myst-spec-ext/src/index.ts
@@ -1,1 +1,151 @@
-export * from './types.js';
+import type {
+  Admonition as SpecAdmonition,
+  AlgorithmLine as SpecAlgorithmLine,
+  AnyWidget as SpecAnyWidget,
+  Aside as SpecAside,
+  Block as SpecBlock,
+  CaptionNumber as SpecCaptionNumber,
+  Cite as SpecCite,
+  CiteGroup as SpecCiteGroup,
+  CiteKind as SpecCiteKind,
+  Code as SpecCode,
+  Container as SpecContainer,
+  CrossReference as SpecCrossReference,
+  DefinitionDescription as SpecDefinitionDescription,
+  DefinitionList as SpecDefinitionList,
+  DefinitionTerm as SpecDefinitionTerm,
+  Delete as SpecDelete,
+  Dependency as SpecDependency,
+  Embed as SpecEmbed,
+  FootnoteDefinition as SpecFootnoteDefinition,
+  FootnoteReference as SpecFootnoteReference,
+  Heading as SpecHeading,
+  Iframe as SpecIframe,
+  Image as SpecImage,
+  Include as SpecInclude,
+  IndexEntry as SpecIndexEntry,
+  InlineExpression as SpecInlineExpression,
+  InlineMath as SpecInlineMath,
+  Link as SpecLink,
+  ListItem as SpecListItem,
+  Math as SpecMath,
+  MathGroup as SpecMathGroup,
+  Output as SpecOutput,
+  Outputs as SpecOutputs,
+  Raw as SpecRaw,
+  SiUnit as SpecSiUnit,
+  Smallcaps as SpecSmallcaps,
+  TabItem as SpecTabItem,
+  TableCell as SpecTableCell,
+  TabSet as SpecTabSet,
+  Underline as SpecUnderline,
+  ContentRecord as SpecContentRecord,
+  DocumentHierarchy as SpecDocumentHierarchy,
+  HeadingRecord as SpecHeadingRecord,
+  MystSearchIndex as SpecMystSearchIndex,
+  SearchRecord as SpecSearchRecord,
+  SearchRecordBase as SpecSearchRecordBase,
+} from 'myst-spec';
+
+/** @deprecated Use `Admonition` from `myst-spec` instead */
+export type Admonition = SpecAdmonition;
+/** @deprecated Use `AlgorithmLine` from `myst-spec` instead */
+export type AlgorithmLine = SpecAlgorithmLine;
+/** @deprecated Use `AnyWidget` from `myst-spec` instead */
+export type AnyWidget = SpecAnyWidget;
+/** @deprecated Use `Aside` from `myst-spec` instead */
+export type Aside = SpecAside;
+/** @deprecated Use `Block` from `myst-spec` instead */
+export type Block = SpecBlock;
+/** @deprecated Use `CaptionNumber` from `myst-spec` instead */
+export type CaptionNumber = SpecCaptionNumber;
+/** @deprecated Use `Cite` from `myst-spec` instead */
+export type Cite = SpecCite;
+/** @deprecated Use `CiteGroup` from `myst-spec` instead */
+export type CiteGroup = SpecCiteGroup;
+/** @deprecated Use `CiteKind` from `myst-spec` instead */
+export type CiteKind = SpecCiteKind;
+/** @deprecated Use `Code` from `myst-spec` instead */
+export type Code = SpecCode;
+/** @deprecated Use `Container` from `myst-spec` instead */
+export type Container = SpecContainer;
+/** @deprecated Use `CrossReference` from `myst-spec` instead */
+export type CrossReference = SpecCrossReference;
+/** @deprecated Use `DefinitionDescription` from `myst-spec` instead */
+export type DefinitionDescription = SpecDefinitionDescription;
+/** @deprecated Use `DefinitionList` from `myst-spec` instead */
+export type DefinitionList = SpecDefinitionList;
+/** @deprecated Use `DefinitionTerm` from `myst-spec` instead */
+export type DefinitionTerm = SpecDefinitionTerm;
+/** @deprecated Use `Delete` from `myst-spec` instead */
+export type Delete = SpecDelete;
+/** @deprecated Use `Dependency` from `myst-spec` instead */
+export type Dependency = SpecDependency;
+/** @deprecated Use `Embed` from `myst-spec` instead */
+export type Embed = SpecEmbed;
+/** @deprecated Use `FootnoteDefinition` from `myst-spec` instead */
+export type FootnoteDefinition = SpecFootnoteDefinition;
+/** @deprecated Use `FootnoteReference` from `myst-spec` instead */
+export type FootnoteReference = SpecFootnoteReference;
+/** @deprecated Use `Heading` from `myst-spec` instead */
+export type Heading = SpecHeading;
+/** @deprecated Use `Iframe` from `myst-spec` instead */
+export type Iframe = SpecIframe;
+/** @deprecated Use `Image` from `myst-spec` instead */
+export type Image = SpecImage;
+/** @deprecated Use `Include` from `myst-spec` instead */
+export type Include = SpecInclude;
+/** @deprecated Use `IndexEntry` from `myst-spec` instead */
+export type IndexEntry = SpecIndexEntry;
+/** @deprecated Use `InlineExpression` from `myst-spec` instead */
+export type InlineExpression = SpecInlineExpression;
+/** @deprecated Use `InlineMath` from `myst-spec` instead */
+export type InlineMath = SpecInlineMath;
+/** @deprecated Use `Link` from `myst-spec` instead */
+export type Link = SpecLink;
+/** @deprecated Use `ListItem` from `myst-spec` instead */
+export type ListItem = SpecListItem;
+/** @deprecated Use `Math` from `myst-spec` instead */
+export type Math = SpecMath;
+/** @deprecated Use `MathGroup` from `myst-spec` instead */
+export type MathGroup = SpecMathGroup;
+/** @deprecated Use `Output` from `myst-spec` instead */
+export type Output = SpecOutput;
+/** @deprecated Use `Outputs` from `myst-spec` instead */
+export type Outputs = SpecOutputs;
+/** @deprecated Use `Raw` from `myst-spec` instead */
+export type Raw = SpecRaw;
+/** @deprecated Use `SiUnit` from `myst-spec` instead */
+export type SiUnit = SpecSiUnit;
+/** @deprecated Use `Smallcaps` from `myst-spec` instead */
+export type Smallcaps = SpecSmallcaps;
+/** @deprecated Use `TabItem` from `myst-spec` instead */
+export type TabItem = SpecTabItem;
+/** @deprecated Use `TableCell` from `myst-spec` instead */
+export type TableCell = SpecTableCell;
+/** @deprecated Use `TabSet` from `myst-spec` instead */
+export type TabSet = SpecTabSet;
+/** @deprecated Use `Underline` from `myst-spec` instead */
+export type Underline = SpecUnderline;
+/** @deprecated Use `ContentRecord` from `myst-spec` instead */
+export type ContentRecord = SpecContentRecord;
+/** @deprecated Use `DocumentHierarchy` from `myst-spec` instead */
+export type DocumentHierarchy = SpecDocumentHierarchy;
+/** @deprecated Use `HeadingRecord` from `myst-spec` instead */
+export type HeadingRecord = SpecHeadingRecord;
+/** @deprecated Use `MystSearchIndex` from `myst-spec` instead */
+export type MystSearchIndex = SpecMystSearchIndex;
+/** @deprecated Use `SearchRecord` from `myst-spec` instead */
+export type SearchRecord = SpecSearchRecord;
+/** @deprecated Use `SearchRecordBase` from `myst-spec` instead */
+export type SearchRecordBase = SpecSearchRecordBase;
+/**
+ * @deprecated Use the export from `myst-spec`
+ *
+ * This is a duplicate declaration, due to esbuild/bun problems
+ */
+export enum SourceFileKind {
+  Article = 'Article',
+  Notebook = 'Notebook',
+  Part = 'Part',
+}
