+++ packages/myst-spec/src/ext.ts
@@ -19,7 +19,7 @@ import type {
   Node,
   CrossReference as SpecCrossReference,
   Link as SpecLink,
-} from 'myst-spec';
+} from './schema.d.ts';
 
 type Visibility = 'show' | 'hide' | 'remove';
 
@@ -313,52 +313,4 @@ export type AnyWidget = {
   css?: string;
   /** Tailwind classes to apply to the container element */
   class?: string;
-};
-// Search types
-
-/**
- * Hierarchy of headers in a document
- */
-export type DocumentHierarchy = {
-  lvl1?: string;
-  lvl2?: string;
-  lvl3?: string;
-  lvl4?: string;
-  lvl5?: string;
-  lvl6?: string;
-};
-
-/**
- * Base type for search records
- */
-export type SearchRecordBase = {
-  hierarchy: DocumentHierarchy;
-  url: string;
-
-  position: number;
-};
-
-/**
- * Record pertaining to section headings
- */
-export type HeadingRecord = SearchRecordBase & {
-  type: keyof DocumentHierarchy;
-};
-
-/**
- * Record pertaining to section content
- */
-export type ContentRecord = SearchRecordBase & {
-  type: 'content';
-  content: string;
-};
-
-/**
- * Indexed search record type
- */
-export type SearchRecord = HeadingRecord | ContentRecord;
-
-export type MystSearchIndex = {
-  version: '1';
-  records: SearchRecord[];
 };
